[FIX] Coding style error (too many function), moved to remove_job file
