+++ src/job_control/remove_job.c
@@ -12,6 +12,28 @@
 #include "types.h"
 #include "jobs.h"
 
+bool compare_pid(void *list_elem, void *compare)
+{
+    job_t *old = list_elem;
+    int status = 0;
+
+    (void)compare;
+    if (waitpid(old->pid, &status, WNOHANG)) {
+        printf("[%d] Done -> %s\n", old->nb_job, old->command);
+        return (true);
+    } else {
+        return (false);
+    }
+}
+
+void free_pid(void *elem)
+{
+    job_t *job = elem;
+
+    free(job->command);
+    free(job);
+}
+
 char **remove_incorrect_char(char **input)
 {
     for (int i = 0; input[i] != NULL; i++) {
